backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
groq==0.4.2
tavily>=1.0.0
httpx==0.27.0
beautifulsoup4==4.12.3
Pillow>=10.0.0